+++ src/web-workbench/classes/logger/Console.js
@@ -1,6 +1,6 @@
 import { ILogger } from '../Logger';
-import { Table } from '@/web-workbench/utils/console';
-import { isStringValue, cleanString } from '@/web-workbench/utils/helper';
+import { Table } from '../..//utils/console';
+import { isStringValue, cleanString } from '../../utils/helper';
 
 export default class ConsoleLogger extends ILogger {
   #onAdd;